pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clearquote-nl2sql"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "sqlglot>=27.0.0,<29.0.0",
  "fastapi==0.115.6",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "sqlalchemy==2.0.36",
  "pandas==2.2.3",
  "openpyxl==3.1.5",
  "python-dotenv==1.0.1",
  "streamlit==1.38.0",
  "httpx==0.27.2",
  "openai>=1.40.0",
  "psycopg[binary]==3.2.3",
  "pytest==8.3.2"
]

[tool.setuptools]
packages = ["app"]
